+++ t/debug.t
@@ -4,15 +4,14 @@
 # Make sure we don't have debugging things left in
 # By J. Stuart McMurray
 # Created 20250216
-# Last Modified 20250302
+# Last Modified 20250830
 
-#set -euo pipefail
-set -eo pipefail
+set -euo pipefail
 
 . t/shmore.subr
 . t/t.subr
 
-tap_plan 3
+tap_plan 2
 
 # Make sure we're not using MQD
 GOT=$(gorun </dev/null 2>&1 | egrep 'MQD DEBUG PACKAGE LOADED$' ||:)
@@ -32,13 +31,5 @@ for i in `jot 6`; do read -p; done
 exec 9>&p; exec 9>&-
 wait
 tap_is "$(cat <"$ERRF")" "" "Nothing on stderr at startup" "$0" $LINENO
-
-# Make sure we didn't leave any stray DEBUGs lying about.
-GOT="$(egrep -InR '#[[:space:]]*DEBUG' | sort -u ||:)"
-tap_is \
-        "$GOT" \
-        "" \
-        "No files with DEBUG comments" \
-        "$0" $LINENO
 
 # vim: ft=sh